+++ projects/ngx-web-tour/src/public-api.ts
@@ -3,8 +3,10 @@
  */
 
 export * from './lib/services/tour.service';
+export * from './lib/services/tour-i18n.service';
 export * from './lib/directives/tour.directive';
 export * from './lib/components/tour-tooltip/tour-tooltip.component';
 export * from './lib/components/tour-backdrop/tour-backdrop.component';
 export * from './lib/models/tour-step.model';
+export * from './lib/models/tour-i18n.model';
 export * from './lib/ngx-web-tour.module';